Updated segment selector o make remesh accessible from python minor update to genesis script generation
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@1074 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/genesis/GenesisFileManager.java
@@ -1480,7 +1480,7 @@ public class GenesisFileManager
         //response.append("tstop = "+project.simulationParameters.duration+"\n");
 
         response.append("setclock 0 "
-                        + convertNeuroConstructTime(project.simulationParameters.getDt()));
+                        + (float)convertNeuroConstructTime(project.simulationParameters.getDt()));
         if (project.genesisSettings.isGenerateComments())
             response.append(" // " + getTimeUnitString());
         response.append("\n");
@@ -1488,7 +1488,7 @@ public class GenesisFileManager
         if (mooseCompatMode())
         {
             response.append("setclock 1 "
-                            + convertNeuroConstructTime(project.simulationParameters.getDt()));
+                            + (float)convertNeuroConstructTime(project.simulationParameters.getDt()));
 
             if (project.genesisSettings.isGenerateComments())
                 response.append(" // " + getTimeUnitString());
